monitor/server/requirements.txt
bottle
grpcio
grpcio-tools